requirements.txt
pyautogui
pytesseract
pillow
pynput
PyQt5
